requirements.txt
appdirs==1.4.3
coverage==4.4
Django==1.11.1
django-nose==1.4.4
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
nose==1.3.7
packaging==16.8
PyJWT==1.5.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
